pyproject.toml
[project]
name = "ecommerce-agents-on-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.22",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
